Refactoring to avoid circular import
