+++ inventory/api/types/item/__init__.py
@@ -1,6 +1,4 @@
-from .inputs import *
 from .item_type import ItemType
-from .payload_types import ItemCreatePayload, ItemDeactivatePayload, ItemUpdatePayload, ItemActivatePayload
 from .user_error_types import SKUNotUniqueError, BarcodeNotUniqueError, ItemNotExistError, ItemIsNotActiveError, \
     ItemIsActiveError, ItemAlreadyHasDocument
